Override Gradle home user to cache dependencies in a Docker layer
In the base Gradle image, the default Gradle home user is a Docker volume. Hence the dependencies, which are written in this volume, are not persisted in a Docker layer. Each Docker build then needs to redownload all the dependencies every time, which is a waste of time and network bandwidth. Given that the dependencies are not likely to change as much as the source code, having them cached in a Docker layer speeds up significantly the further Docker builds and the overall development feedback loop. To do so, we copy first copy the build.gradle and settings.gradle files and run a build which will download all the dependencies. In a next step we then copy the source code and run a Gradle build again, which will make use of the dependencies downloaded previously. When the source code changes, the Docker build can reuse the cache up to the point where the dependencies were downloaded, and thus just needs to build the source code without having to redownload all the dependencies again.